adds initial design for search page
